refactor(dev): extract dev server logic from cli commands
