+++ packages/remix-dev/compiler/index.ts
@@ -1,5 +1,5 @@
 export { build } from "./build";
-export { watch } from "./watch";
+export { type WatchOptions, watch } from "./watch";
 
 export { type CompileOptions, parseMode } from "./options";
 export { logCompileFailure } from "./onCompileFailure";
